+++ e2e/validation/supportedInvalidFails.spec.ts
@@ -13,6 +13,8 @@ test('an incomplete record drawn together from filled in sections does not valid
 
   // DON'T CHECK element with id 'topic-living_and_working_in_antarctica'
 
+  // Additional properties aren't set as it's not needed to test validation failure
+
   // click the button with id 'validation-use-current'
   await page.click('button#validation-use-current')